requirements.txt
watchdog==3.0.0
requests==2.31.0
websocket-client==1.6.4